Move actors and actor groups from dashboard to context
